requirements.txt
imap-tools
pandas
openpyxl
python-dotenv
python-dateutil